Refactor helper methods to go inside Task, making hook_manager an optional argument, and adding parallel as boolean flag
